requirements.txt
aio_pika==9.4.2
attrs==22.2.0
colorama==0.4.6
ConfigParser==7.0.0
cryptography==42.0.8
docutils==0.20.1
fastapi==0.111.1
filelock==3.15.4
HTMLParser==0.0.2
importlib_metadata==6.8.0
ipython==8.12.3
ipywidgets==8.1.3
jnius==1.1.0
keyring==24.3.0
loguru==0.7.2
passlib==1.7.4
Pillow==10.1.0
Pillow==10.4.0
protobuf==5.27.3
psycopg2==2.8.6
pydantic==2.8.2
pyOpenSSL==24.1.0
pyOpenSSL==24.2.1
pyotp==2.9.0
python_jose==3.3.0
redis==4.4.2
Sphinx==8.0.2
SQLAlchemy==2.0.31
starlette==0.38.2
thread==2.0.5
tornado==6.4
urllib3_secure_extra==0.1.0
uvicorn==0.30.4
xmlrpclib==1.0.1
